+++ time/DateTime.js
@@ -20,9 +20,9 @@ const addExamples = require('../_util/addExamples')
 
 const DateTime = Joi.string()
   .isoDate()
-  .pattern(/^\d{4}-(0[1-9]|1[0-2])-((0[1-9]|[1-2]\d)|30|31)T(0\d|1\d|2[0-3]):([0-5]\d):([0-5]\d)\.\d{3}Z$/)
-  .description("Moment in time, as ISO-8601 expressed in UTC ('Z'), to ms precision.")
+  .pattern(/^\d{4}-(0[1-9]|1[0-2])-((0[1-9]|[1-2]\d)|30|31)T(0\d|1\d|2[0-3]):([0-5]\d):([0-5]\d)\.\d{3,}Z$/)
+  .description("Moment in time, as ISO-8601 expressed in UTC ('Z'), to ms precision or more precise.")
 
-const examples = ['2020-01-23T15:22:39.212Z']
+const examples = ['2020-01-23T15:22:39.212Z', '2020-01-23T15:22:39.21254888Z']
 
 module.exports = { dateTimeExamples: examples, DateTime: addExamples(DateTime, examples) }